refactored: raw_connection_interface, secure_conn_interface and stream_muxer abc.py
