+++ libp2p/custom_types.py
@@ -8,8 +8,8 @@ from typing import (
 )
 
 if TYPE_CHECKING:
-    from libp2p.network.stream.net_stream_interface import INetStream  # noqa: F401
-    from libp2p.stream_muxer.abc import IMuxedStream  # noqa: F401
+    from libp2p.abc import IMuxedStream  # noqa: F401
+    from libp2p.abc import INetStream  # noqa: F401
 
 TProtocol = NewType("TProtocol", str)
 StreamHandlerFn = Callable[["INetStream"], Awaitable[None]]
