+++ libp2p/host/ping.py
@@ -6,6 +6,7 @@ import trio
 
 from libp2p.abc import (
     IHost,
+    INetStream,
 )
 from libp2p.custom_types import (
     TProtocol,
@@ -14,9 +15,6 @@ from libp2p.network.stream.exceptions import (
     StreamClosed,
     StreamEOF,
     StreamReset,
-)
-from libp2p.network.stream.net_stream_interface import (
-    INetStream,
 )
 from libp2p.peer.id import ID as PeerID
