--- libp2p/network/connection/raw_connection_interface.py
@@ -1,9 +1,0 @@
-from libp2p.io.abc import (
-    ReadWriteCloser,
-)
-
-
-class IRawConnection(ReadWriteCloser):
-    """A Raw Connection provides a Reader and a Writer."""
-
-    is_initiator: bool
