+++ libp2p/network/connection/swarm_connection.py
@@ -5,14 +5,12 @@ from typing import (
 import trio
 
 from libp2p.abc import (
+    IMuxedConn,
+    IMuxedStream,
     INetConn,
 )
 from libp2p.network.stream.net_stream import (
     NetStream,
-)
-from libp2p.stream_muxer.abc import (
-    IMuxedConn,
-    IMuxedStream,
 )
 from libp2p.stream_muxer.exceptions import (
     MuxedConnUnavailable,
