--- libp2p/network/stream/net_stream_interface.py
@@ -1,33 +1,0 @@
-from abc import (
-    abstractmethod,
-)
-
-from libp2p.custom_types import (
-    TProtocol,
-)
-from libp2p.io.abc import (
-    ReadWriteCloser,
-)
-from libp2p.stream_muxer.abc import (
-    IMuxedConn,
-)
-
-
-class INetStream(ReadWriteCloser):
-    muxed_conn: IMuxedConn
-
-    @abstractmethod
-    def get_protocol(self) -> TProtocol:
-        """
-        :return: protocol id that stream runs on
-        """
-
-    @abstractmethod
-    def set_protocol(self, protocol_id: TProtocol) -> None:
-        """
-        :param protocol_id: protocol id that stream runs on
-        """
-
-    @abstractmethod
-    async def reset(self) -> None:
-        """Close both ends of the stream."""
