+++ libp2p/security/insecure/transport.py
@@ -1,3 +1,7 @@
+from libp2p.abc import (
+    IRawConnection,
+    ISecureConn,
+)
 from libp2p.crypto.exceptions import (
     MissingDeserializerError,
 )
@@ -23,9 +27,6 @@ from libp2p.io.msgio import (
 from libp2p.network.connection.exceptions import (
     RawConnError,
 )
-from libp2p.network.connection.raw_connection_interface import (
-    IRawConnection,
-)
 from libp2p.peer.id import (
     ID,
 )
@@ -37,9 +38,6 @@ from libp2p.security.base_transport import (
 )
 from libp2p.security.exceptions import (
     HandshakeFailure,
-)
-from libp2p.security.secure_conn_interface import (
-    ISecureConn,
 )
 
 from .pb import (
