+++ libp2p/security/noise/transport.py
@@ -1,3 +1,7 @@
+from libp2p.abc import (
+    IRawConnection,
+    ISecureConn,
+)
 from libp2p.crypto.keys import (
     KeyPair,
     PrivateKey,
@@ -5,14 +9,8 @@ from libp2p.crypto.keys import (
 from libp2p.custom_types import (
     TProtocol,
 )
-from libp2p.network.connection.raw_connection_interface import (
-    IRawConnection,
-)
 from libp2p.peer.id import (
     ID,
-)
-from libp2p.security.secure_conn_interface import (
-    ISecureConn,
 )
 from libp2p.security.secure_transport_interface import (
     ISecureTransport,
