+++ libp2p/security/security_multistream.py
@@ -5,11 +5,12 @@ from collections import (
     OrderedDict,
 )
 
+from libp2p.abc import (
+    IRawConnection,
+    ISecureConn,
+)
 from libp2p.custom_types import (
     TProtocol,
-)
-from libp2p.network.connection.raw_connection_interface import (
-    IRawConnection,
 )
 from libp2p.peer.id import (
     ID,
@@ -22,9 +23,6 @@ from libp2p.protocol_muxer.multiselect_client import (
 )
 from libp2p.protocol_muxer.multiselect_communicator import (
     MultiselectCommunicator,
-)
-from libp2p.security.secure_conn_interface import (
-    ISecureConn,
 )
 from libp2p.security.secure_transport_interface import (
     ISecureTransport,
