+++ libp2p/stream_muxer/mplex/mplex.py
@@ -5,6 +5,11 @@ from typing import (
 
 import trio
 
+from libp2p.abc import (
+    IMuxedConn,
+    IMuxedStream,
+    ISecureConn,
+)
 from libp2p.custom_types import (
     TProtocol,
 )
@@ -19,13 +24,6 @@ from libp2p.network.connection.exceptions import (
 )
 from libp2p.peer.id import (
     ID,
-)
-from libp2p.security.secure_conn_interface import (
-    ISecureConn,
-)
-from libp2p.stream_muxer.abc import (
-    IMuxedConn,
-    IMuxedStream,
 )
 from libp2p.utils import (
     decode_uvarint_from_stream,
