+++ libp2p/tools/factories.py
@@ -24,6 +24,9 @@ from libp2p import (
 )
 from libp2p.abc import (
     IHost,
+    INetStream,
+    IRawConnection,
+    ISecureConn,
 )
 from libp2p.crypto.ed25519 import create_new_key_pair as create_ed25519_key_pair
 from libp2p.crypto.keys import (
@@ -46,14 +49,8 @@ from libp2p.io.abc import (
 from libp2p.network.connection.raw_connection import (
     RawConnection,
 )
-from libp2p.network.connection.raw_connection_interface import (
-    IRawConnection,
-)
 from libp2p.network.connection.swarm_connection import (
     SwarmConn,
-)
-from libp2p.network.stream.net_stream_interface import (
-    INetStream,
 )
 from libp2p.network.swarm import (
     Swarm,
@@ -95,9 +92,6 @@ from libp2p.security.noise.messages import (
 from libp2p.security.noise.transport import PROTOCOL_ID as NOISE_PROTOCOL_ID
 from libp2p.security.noise.transport import Transport as NoiseTransport
 import libp2p.security.secio.transport as secio
-from libp2p.security.secure_conn_interface import (
-    ISecureConn,
-)
 from libp2p.security.secure_transport_interface import (
     ISecureTransport,
 )
